fix(server): bounds-check search limit env parsing
CodeQL flagged the two new uint32(envconfig.Int(...)) conversions in the
SearchConfig block as go/incorrect-integer-conversion (CWE-190): an int
parsed from the environment was narrowed to uint32 without an upper-bound
check.

Add envconfig.Uint32, which parses with strconv.ParseUint(.., 10, 32) so an
out-of-range value errors out to the default instead of silently truncating,
and switch LANTERN_SEARCH_DEFAULT_LIMIT / LANTERN_SEARCH_MAX_LIMIT to it.

+++ server/provider/provider.go
@@ -221,8 +221,8 @@ func NewConfig() *Config {
 		},
 		Search: SearchConfig{
 			Enabled:      envconfig.Bool("LANTERN_SEARCH_ENABLED", true),
-			DefaultLimit: uint32(envconfig.Int("LANTERN_SEARCH_DEFAULT_LIMIT", 100)),
-			MaxLimit:     uint32(envconfig.Int("LANTERN_SEARCH_MAX_LIMIT", 1000)),
+			DefaultLimit: envconfig.Uint32("LANTERN_SEARCH_DEFAULT_LIMIT", 100),
+			MaxLimit:     envconfig.Uint32("LANTERN_SEARCH_MAX_LIMIT", 1000),
 		},
 		MutationLog: loadMutationLogConfig(),
 		Replication: loadReplicationConfig(),
